+++ emtorch/subtasks/files.py
@@ -7,13 +7,11 @@ Module holding sub-tasks related to files.
 """
 
 import logging
-import os
 from typing import Self
 
 from ..config import Config
-from ..context import CaseContext, Context
+from ..context import CaseContext
 from ..context.template import Template
-from ..io import IOLoop
 from ..io.streams import StreamWriter
 from .subtask import BasicSubTask
 
@@ -27,8 +25,6 @@ class FileWriter(BasicSubTask):
         append: bool,
         lines: list[Template],
         encoding: str,
-        timeout: float,
-        io: IOLoop,
     ):
         super().__init__(name, logging.getLogger(__name__))
 
@@ -36,48 +32,37 @@ class FileWriter(BasicSubTask):
         self._lines = lines
         self._append = append
         self._encoding = encoding
-        self._timeout = timeout
-        self._io = io
         self._writer: None | StreamWriter = None
 
+        self._context: CaseContext | None = None
+
     def basic_start(self, context: CaseContext) -> bool:
-        lines = [line.evaluate(context) for line in self._lines]
-        data = "\n".join(lines).encode(self._encoding)
-
-        path = self._path.evaluate(context)
-        mode = ("a" if self._append else "w") + "b"
-        flags = os.O_WRONLY | os.O_CREAT | (os.O_APPEND if self._append else os.O_TRUNC)
-        try:
-            fd = os.open(path, flags, 0o666)
-            os.set_blocking(fd, False)
-            file = os.fdopen(fd, mode)
-        except IOError as ex:
-            self.logger.error(f"Open file failed '{path}' - {ex}")
-            return False
-
-        self._writer = StreamWriter(self.name, file, data)
-        self._io.register(self._writer)
-
+        self._context = context
         return True
 
     def finish(self) -> BasicSubTask.Result:
-        assert self._writer
+        assert self._context
 
-        if self._writer.wait_for_done(self._timeout):
-            self.logger.info(f"File written: {self._writer.name()}")
-            return self.Result.SUCCESS
+        lines = [line.evaluate(self._context) for line in self._lines]
+        data = "\n".join(lines).encode(self._encoding)
 
-        self.logger.error(f"File write timeout: {self._writer.name()}")
-        return self.Result.TIMEOUT
+        path = self._path.evaluate(self._context)
+
+        try:
+            with open(path, "wb") as file:
+                file.write(data)
+        except IOError as ex:
+            self.logger.error(f"File write error: {ex}")
+            return self.Result.ERROR
+
+        return self.Result.SUCCESS
 
     @classmethod
-    def from_config(cls, name: str, config: Config, context: Context) -> Self:
+    def from_config(cls, name: str, config: Config) -> Self:
         return cls(
             name=name,
             path=Template(config.get_str("path")),
             append=config.get_bool("append", fallback=False),
             lines=[Template(line) for line in config.get_str_list("lines")],
             encoding=config.get_str("encoding", fallback="utf-8"),
-            timeout=config.get_float("timeout", fallback=5),
-            io=context.worker(IOLoop),
         )